Increased the range of IP addresses.
Increased the range of IP addresses available to help prevent filtration. The logic is that, for each greater number of IP addresses used, and assuming hidden form IP fields can still be written to, the ratio of false-positives marked by the website risks being roughly equal to the number of authentic reports declined as being fake. This doesn't preclude forensics to narrow down IP addresses, but if deployed in Texas or through Texas proxies then countering this program might require more time and money than running the site ever was.